Refactor : StartButton.tsx 파일 위치를 팀 규칙에 맞게 이동
